+++ backend/backend/Domain/VentaModel.cs
@@ -5,14 +5,14 @@
         public int ID { get; set; }
         public int IDCliente { get; set; }
         public DateTime Fecha { get; set; }
-        public int Semana { get; set; }
+        public byte Semana { get; set; }
         public int IDCafe { get; set; }
         public int Cantidad { get; set; }
-        public double PrecioTotal { get; set; }
-        public double Pago { get; set; }
-        public double Saldo { get; set; }
+        public decimal PrecioTotal { get; set; }
+        public decimal Pago { get; set; }
+        public decimal Saldo { get; set; }
         public bool Debe { get; set; }
-        public double PesoTotal { get; set; }
+        public decimal PesoTotal { get; set; }
 
 
     }
